Split DCR/DTVR implementations
- inconvenient to maintain a single multipurpose codebase
- some configuration/signature breaking changes
