+++ .changeset/lemon-hotels-shop.md
@@ -1,0 +1,8 @@
+---
+"@theoplayer/nielsen-connector-web": major
+---
+
+Split DCR/DTVR implementations
+
+- inconvenient to maintain a single multipurpose codebase
+- some configuration/signature breaking changes
